LPS-194420 refactor KBArticleAttachmentKBUploadFileEntryHandler.java
moving class to UploadKBArticleAttachmentsMVCActionCommand.java and deleting handler class
